// apps/frontend/src/api/generated/admin/model/customer.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  deletedAt: string | null;
}